pyproject.toml
[project]
name = "fuz"
version = "0.1.0"
description = "Minimal Django ORM-based CLI tool"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.0",
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "hdbscan>=0.8.40",
    "ipython>=9.6.0",
    "pgvector>=0.4.1",
    "psycopg[binary]>=3.2.10",
    "pyinstaller>=6.16.0",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.1",
]

[project.scripts]
fuz = "fuz.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
